EWPP-421: Specify the #empty_value to allow AJAX operations on the form.
Config entity forms build the entity copying all the values submitted. If an AJAX callback rebuilds
the form, the empty option value has been set to the entity. Due to the order of form building
operations, an invalid option error will be thrown as the #default_value specified in the form
doesn't exist.